Moves AbstractChessEvaluator & AbstractPieceSquareTable + update action
