+++ src/main/java/com/fathzer/chess/utils/evaluators/simplified/EndGameKingSquareTable.java
@@ -1,6 +1,6 @@
 package com.fathzer.chess.utils.evaluators.simplified;
 
-import com.fathzer.chess.utils.evaluators.utils.AbstractPieceSquareTable;
+import com.fathzer.chess.utils.evaluators.AbstractPieceSquareTable;
 
 class EndGameKingSquareTable extends AbstractPieceSquareTable {
 	private static final int[] EVAL = new int[] {
